revert refactor for circ dep
