+++ clients/fides-js/src/lib/initialize.ts
@@ -7,6 +7,7 @@ import {
   CookieIdentity,
   CookieKeyConsent,
   CookieMeta,
+  FidesCookie,
   getCookieByName,
   getOrMakeFidesCookie,
   isNewFidesCookie,
@@ -24,7 +25,6 @@ import {
   PrivacyExperience,
   SaveConsentPreference,
   UserGeolocation,
-  FidesCookie,
 } from "./consent-types";
 import {
   constructFidesRegionString,
