+++ clients/fides-js/src/lib/preferences.ts
@@ -1,6 +1,5 @@
 import {
   ConsentMethod,
-  FidesCookie,
   FidesOptions,
   PrivacyExperience,
   PrivacyPreferencesRequest,
@@ -8,7 +7,11 @@ import {
   UserConsentPreference,
 } from "./consent-types";
 import { debugLog } from "./consent-utils";
-import { removeCookiesFromBrowser, saveFidesCookie } from "./cookie";
+import {
+  FidesCookie,
+  removeCookiesFromBrowser,
+  saveFidesCookie,
+} from "./cookie";
 import { dispatchFidesEvent, FidesEventExtraDetails } from "./events";
 import { patchUserPreference } from "../services/api";
 import { TcfSavePreferences } from "./tcf/types";
