+++ clients/privacy-center/components/consent/NoticeDrivenConsent.tsx
@@ -11,7 +11,7 @@ import {
   PrivacyNotice,
   ConsentContext,
   FidesCookie,
-  PrivacyNoticeWithPreference,
+  PrivacyNoticeExtended,
 } from "fides-js";
 import { useAppSelector } from "~/app/hooks";
 import {
@@ -55,12 +55,12 @@ export const resolveConsentValue = (
   const gpcEnabled =
     !!notice.has_gpc_flag &&
     context.globalPrivacyControl === true &&
-    !noticeHasConsentInCookie(notice as PrivacyNoticeWithPreference, cookie);
+    !noticeHasConsentInCookie(notice as PrivacyNoticeExtended, cookie);
   if (gpcEnabled) {
     return UserConsentPreference.OPT_OUT;
   }
   const preferenceExistsInCookie = noticeHasConsentInCookie(
-    notice as PrivacyNoticeWithPreference,
+    notice as PrivacyNoticeExtended,
     cookie
   );
   if (preferenceExistsInCookie) {
