+++ clients/privacy-center/cypress/e2e/consent-banner.cy.ts
@@ -1857,7 +1857,6 @@ describe("Consent overlay", () => {
         experience: {
           privacy_notices: [
             mockPrivacyNotice({
-              // fixme- even if consent default pref is opt-in, GPC should apply
               name: "Applied",
               notice_key: "applied",
               has_gpc_flag: true,
